# openapi/protocol-v1.yaml
openapi: 3.1.0
info:
  title: unsea Protocol API v1
  version: 1.0.0-draft
  description: |
    新一代 WhatsApp 协议层接口（基于 Baileys 7.x）。
    本规范覆盖《Baileys 协议层接口封装与 2000 账号承载方案》§ 3 + 附录 A。

    ## 设计原则

    - 协议层只负责 ws / Noise / creds / 重连，业务编排在 Java 业务层
    - 所有状态查询都带 evidence（时间戳 + 来源）
    - 所有重连相关都对应明确的语义码（PROXY_FAILED / RATE_LIMITED / NEED_REAUTH ...）
    - 接口契约稳定，protobuf 内部码不外泄

    ---

    ## 业务场景导航

    ### 🔑 账号登录 / 授权（3 种方式 + 4 种导入）

    | 业务场景 | 接口 | 说明 |
    |---|---|---|
    | 首次绑定新号（pairing code） | `POST /v1/auth/pairing-code` | 用户在手机 WA "已登录设备" 里输入 8 位码；90s 不绑会自动释放 |
    | 首次绑定新号（扫码） | `POST /v1/auth/qrcode` | 用户用手机 WA 扫返回的二维码；180s 不绑会自动释放 |
    | 标准 Baileys JSON 导入（推荐） | `POST /v1/accounts/import/baileys-json` | 完整 creds + keys，导入即可上线 |
    | 全参数登录（30+ 字段） | `POST /v1/accounts/import/params` | 移植自 Apifox `paramsLogin` |
    | 六段参数登录 | `POST /v1/accounts/import/six` | 移植自 Apifox `sixLogin`，含 deviceIdentityKey |
    | 旧 JSON 登录 | `POST /v1/accounts/import/legacy-json` | 移植自 Apifox `jsonLogin`，base64 |
    | 批量导入（最多 500） | `POST /v1/accounts/import/batch` | 含冷启动节流，避免一次性灌爆 worker |
    | 短信注册新号 | （不支持） | WA 业务不开放此能力 |

    **关键概念**：换 IP 不需要重新授权（详见 § 4.3）。一次 pairing 后 creds 永久持有，重连用本地 creds 即可。

    ### 🚪 账号生命周期

    | 业务场景 | 接口 |
    |---|---|
    | 已绑账号上线（用已存 creds） | `POST /v1/accounts/{accountId}/online` |
    | 主动下线（保留 creds，可重新 online） | `POST /v1/accounts/{accountId}/offline` |
    | 退出并移除设备 | `POST /v1/accounts/{accountId}/logout` |

    ### 🏷️ 账号状态查询

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/contacts/{jid}/save:
    post: { tags: [contacts], summary: "[P2] 添加/编辑联系人", operationId: contactSave, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/contacts/{jid}:
    delete: { tags: [contacts], summary: "[P2] 删除联系人", operationId: contactDelete, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/contacts/{jid}/block:
    post: { tags: [contacts], summary: "[P2] 拉黑", operationId: contactBlock, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/contacts/{jid}/unblock:
    post: { tags: [contacts], summary: "[P2] 取消拉黑", operationId: contactUnblock, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/chats/{jid}/mute:
    post: { tags: [contacts], summary: "[P2] 静音", operationId: chatMute, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/chats/{jid}/clear:
    post: { tags: [contacts], summary: "[P2] 清空聊天", operationId: chatClear, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/chats/{jid}/delete:
    post: { tags: [contacts], summary: "[P2] 删除会话", operationId: chatDelete, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/chats/{jid}/archive:
    post: { tags: [contacts], summary: "[P2] 归档", operationId: chatArchive, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/chats/{jid}/pin:
    post: { tags: [contacts], summary: "[P2] 置顶", operationId: chatPin, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /v1/chats/{jid}/mark-read:
    post: { tags: [contacts], summary: "[P2] 标记已读", operationId: chatMarkRead, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }

  # ──────────────────── A.8/A.9 资料 + Business（占位） ────────────────────
...
  /v1/profile/business/{jid}:
    get:
      tags: [business]
      summary: 获取 Business profile（catalog/hours/category/address）
      description: |
        Baileys `sock.getBusinessProfile(jid)`.
        **如果 jid 对应账号不是 Business 号，返回 null（200，body 为空对象）**。
        业务侧可以用此判断他人账号是否 Business：
          if (response.profile == null) → 个人号
          if (response.profile != null) → Business 号
      operationId: businessProfileGet
      parameters:
        - in: path
          name: jid
          required: true
          schema: { type: string }
        - in: query
          name: accountId
          required: true
          description: 查询者账号
          schema: { type: string }
      responses:
        '200':
          description: Business profile 或 null
          content:
            application/json:
              schema:
                type: object
                properties:
                  jid: { type: string }
                  isBusiness: { type: boolean }
                  profile:
                    type: object
                    nullable: true
                    description: 个人号此字段为 null
                    properties:
                      wid: { type: string }
                      description: { type: string, nullable: true }
                      website:
                        type: array
                        items: { type: string }
                      email: { type: string, nullable: true }
                      address: { type: string, nullable: true }
                      category: { type: string, nullable: true }
                      businessHours:
                        type: object
                        nullable: true
                        properties:
                          timezone: { type: string }
                          config:
                            type: array
                            items:
                              type: object
                              properties:
                                day_of_week: { type: string }
                                mode: { type: string, enum: [open_24h, closed, specific_hours] }
                                open_time: { type: integer, nullable: true }
                                close_time: { type: integer, nullable: true }
  /v1/profile/business/update:
    post:
      tags: [business]
      summary: 更新自己的 Business profile
      description: |
        Baileys `updateBussinesProfile(...)` （注意 Baileys 拼写：Bussines）。
        **必须是 Business 号才能调**。个人号调用返回 422。
      operationId: businessProfileUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BusinessProfileUpdateBody' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '422':
          description: 账号不是 Business 号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/business/catalog:
    get:
      tags: [business]
      summary: 获取 Business catalog（商品列表）
      description: Baileys `getCatalog({ jid, limit, cursor })`
      operationId: businessCatalog
      parameters:
        - in: query
          name: accountId
          required: true
          schema: { type: string }
        - in: query
          name: jid
          required: true
          description: 目标 Business 账号 jid（可以是自己也可以是其他 Business）
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: cursor
          description: 上次返回的 nextCursor，首次留空
          schema: { type: string }
      responses:
        '200':
          description: 商品列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogResult' }
        '422':
          description: 目标账号不是 Business 号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/business/product:
    post:
      tags: [business]
      summary: 创建 / 更新 / 删除商品
      description: |
        Baileys `productCreate / productUpdate / productDelete`，按 action 分流。
        **必须是 Business 号才能调**。
      operationId: businessProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductMutationBody' }
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductMutationResult' }
        '422':
          description: 账号不是 Business 号
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ──────────────────── A.10 频道 newsletter（P3 完整 schema） ────────────────────
  /v1/channels/follow:
    post:
      tags: [channels]
      summary: 订阅频道
      description: Baileys `newsletterFollow(jid)`
      operationId: channelFollow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChannelOpBody' }
      responses:
        '200':
          description: 已订阅
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  jid: { type: string }
  /v1/channels/unfollow:
    post:
      tags: [channels]
      summary: 取消订阅频道
      description: Baileys `newsletterUnfollow(jid)`
      operationId: channelUnfollow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChannelOpBody' }
      responses:
        '200':
          description: 已取消订阅
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  jid: { type: string }
  /v1/channels/{jid}/messages:
    get:
      tags: [channels]
      summary: 获取频道历史消息
      description: Baileys `newsletterFetchMessages(jid, count, since, after)`
      operationId: channelMessages
      parameters:
        - in: path
          name: jid
          required: true
          schema: { type: string }
        - in: query
          name: accountId
          required: true
          schema: { type: string }
        - in: query
          name: count
          description: 拉取条数
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: since
          description: unix 时间戳（秒），仅返回此时间之后的消息
          schema: { type: integer }
        - in: query
          name: after
          description: 频道内部 serverId 游标
          schema: { type: string }
      responses:
        '200':
          description: 频道消息列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChannelMessageList' }
  /v1/channels/{jid}/reaction:
    post:
      tags: [channels]
      summary: 给频道消息加 / 移除表情反应
      description: Baileys `newsletterReactMessage(jid, serverId, reaction)`
      operationId: channelReaction
      parameters:
        - in: path
          name: jid
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChannelReactionBody' }
      responses:
        '200':
          description: 反应已发送
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }

  # ──────────────────── Admin 运维 / 诊断（仅内网） ────────────────────
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        evidence: { $ref: '#/components/schemas/Evidence' }
        details:
          type: object
          additionalProperties: true
          description: |
            错误详情。`NOT_OWNER` 时包含 accountId/currentWorkerId/ownerWorkerId/ownerEndpoint，
            业务层应刷新 owner 后直连 ownerEndpoint 重试一次。
    Evidence:
      type: object
      description: 状态来源证据（详见 § 4.7.1 三层证据模型）
      properties:
        wsOpen: { type: boolean }
        connectionField:
          type: string
          enum: [connecting, open, close]
        lastDateRecv: { type: string, format: date-time }
        lastPingAckAt: { type: string, format: date-time }
        ageMs: { type: integer }
        keepAliveIntervalMs: { type: integer, example: 30000 }
        slotReleased:
          type: boolean
          description: true 表示该账号当前不占 worker 在线容量
        reason:
          type: string
          nullable: true
          description: slotReleased 或状态变更原因
        updatedAt:
          type: string
          format: date-time
          description: runtime 状态更新时间
...
    BusinessProfileUpdateBody:
      type: object
      required: [accountId]
      properties:
        accountId: { type: string }
        description: { type: string, maxLength: 256, nullable: true }
        category: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        website:
          type: array
          maxItems: 2
          items: { type: string, format: uri }
        address: { type: string, nullable: true }
        businessHours:
          type: object
          nullable: true
          properties:
            timezone: { type: string, example: "Asia/Shanghai" }
            config:
              type: array
              items:
                type: object
                properties:
                  day_of_week:
                    type: string
                    enum: [mon, tue, wed, thu, fri, sat, sun]
                  mode:
                    type: string
                    enum: [open_24h, closed, specific_hours]
                  open_time:
                    type: integer
                    description: 分钟数（0-1439）
                  close_time:
                    type: integer
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        priceAmount1000:
          type: integer
          description: 价格 × 1000（避免浮点），例如 USD 9.99 = 9990
        currency: { type: string, example: "USD" }
        retailerId: { type: string, nullable: true }
        url: { type: string, format: uri, nullable: true }
        images:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              mimetype: { type: string }
        isHidden: { type: boolean, default: false }
        availability:
          type: string
          enum: [in_stock, out_of_stock]
    CatalogResult:
      type: object
      properties:
        accountId: { type: string }
        targetJid: { type: string }
        total: { type: integer }
        products:
          type: array
          items: { $ref: '#/components/schemas/Product' }
        nextCursor: { type: string, nullable: true }
    ProductMutationBody:
      type: object
      required: [accountId, action]
      properties:
        accountId: { type: string }
        action:
          type: string
          enum: [create, update, delete]
        product:
          $ref: '#/components/schemas/Product'
          description: "create / update 时必填；delete 时只需 productIds"
        productIds:
          type: array
          items: { type: string }
          description: "delete 时使用"
    ProductMutationResult:
      type: object
      properties:
        action: { type: string }
        success: { type: boolean }
        product:
          $ref: '#/components/schemas/Product'
        deletedIds:
          type: array
          items: { type: string }

    # ──── 频道 newsletter ────
    ChannelOpBody:
      type: object
      required: [accountId, jid]
      properties:
        accountId: { type: string }
        jid:
          type: string
          example: "120363xxx@newsletter"
    ChannelReactionBody:
      type: object
      required: [accountId, serverId, reaction]
      properties:
        accountId: { type: string }
        serverId:
          type: string
          description: 频道消息的 serverId（newsletter 消息独有，不是普通 messageId）
        reaction:
          type: string
          description: 单个 emoji；传空字符串 '' 表示移除
    ChannelMessage:
      type: object
      properties:
        serverId: { type: string }
        timestamp: { type: integer }
        senderJid: { type: string }
        type:
          type: string
          enum: [text, image, video, document, audio]
        content: { type: string, nullable: true }
        mediaUrl: { type: string, format: uri, nullable: true }
        reactions:
          type: array
          items:
            type: object
            properties:
              emoji: { type: string }
              count: { type: integer }
    ChannelMessageList:
      type: object
      properties:
        jid: { type: string }
        total: { type: integer }
        messages:
          type: array
          items: { $ref: '#/components/schemas/ChannelMessage' }
        nextCursor: { type: string, nullable: true }


# ──────────────────── webhooks（事件订阅，NATS subject: unsea.v1.events.*） ────────────────────
...